requirements.txt
pandas
psycopg2-binary
openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pdfplumber
python-docx
streamlit-quill
python-dotenv
openpyxl
requests>=2.31.0
streamlit>=1.28.0
google-generativeai
ics
plotly
streamlit-option-menu>=0.3.6

